requirements.txt
# LiveKit Python SDK
livekit-api>=0.11.0
livekit-protocol>=0.11.0

# Web framework for webhook server (optional - can use FastAPI, Flask, etc.)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# For webhook verification (if not using livekit-server-sdk)
pyjwt>=2.8.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

